requirements.txt
Pillow==8.4.0
pytesseract==0.3.8
opencv-python-headless==4.8.1.78
matplotlib==3.4.3